Time: 195 ms (91.57%), Space: 42.8 MB (77.83%) - LeetHub

+++ 3179-find-the-n-th-value-after-k-seconds/3179-find-the-n-th-value-after-k-seconds.java
@@ -1,6 +1,6 @@
 class Solution {
     public int valueAfterKSeconds(int n, int k) {
-        long prefix[] = new long[n];
+        int prefix[] = new int[n];
         Arrays.fill(prefix, 1);
 
         while(k-- > 0){
@@ -9,6 +9,6 @@ class Solution {
             }
         }
 
-        return (int) prefix[n - 1];
+        return  prefix[n - 1];
     }
 }
